+++ ta4j/src/main/java/org/ta4j/core/indicators/CoppockCurveIndicator.java
@@ -36,6 +36,18 @@ public class CoppockCurveIndicator extends CachedIndicator<Decimal> {
     private final WMAIndicator wma;
     
     /**
+      * Constructor with default values: <br/>
+      * - longRoCTimeFrame=14 <br/>
+      * - shortRoCTimeFrame=11 <br/>
+      * - wmaTimeFrame=10
+      * 
+      * @param indicator
+    */
+    public CoppockCurveIndicator(Indicator<Decimal> indicator) {
+        this(indicator, 14, 11, 10);
+    }
+    
+    /**
      * Constructor.
      * @param indicator the indicator (usually close price)
      * @param longRoCTimeFrame the time frame for long term RoC