+++ ta4j/src/main/java/org/ta4j/core/indicators/EMAIndicator.java
@@ -62,4 +62,9 @@ public class EMAIndicator extends RecursiveCachedIndicator<Decimal> {
         Decimal emaPrev = getValue(index - 1);
         return indicator.getValue(index).minus(emaPrev).multipliedBy(multiplier).plus(emaPrev);
     }
+    
+    @Override
+    public String toString() {
+        return getClass().getSimpleName() + " timeFrame: " + timeFrame;
+    }
 }