+++ ta4j/src/main/java/org/ta4j/core/indicators/adx/DirectionalMovementPlusIndicator.java
@@ -53,6 +53,6 @@ public class DirectionalMovementPlusIndicator extends CachedIndicator<Decimal> {
 
     @Override
     public String toString() {
-        return getClass().getSimpleName()+ "timeFrame: "+timeFrame;
+        return getClass().getSimpleName() + "timeFrame: " + timeFrame;
     }
 }